requirements.txt
alembic==1.18.4
amqp==5.3.1
annotated-doc @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_annotated-doc_1762868213/work
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1733247046149/work
anthropic==0.84.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
backports.zstd @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_backports.zstd_1767045020/work
bcrypt==5.0.0
billiard==4.2.4
blis @ file:///Users/runner/miniforge3/conda-bld/cython-blis_1763427432869/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1764017116075/work
build==1.4.0
catalogue @ file:///Users/runner/miniforge3/conda-bld/catalogue_1756301975418/work
celery==5.6.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1772001073725/work/certifi
cffi==2.0.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
chromadb==1.5.2
click @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_click_1764518652/work
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib @ file:///home/conda/feedstock_root/build_artifacts/cloudpathlib-meta_1759961686710/work
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_cloudpickle_1765303462/work
confection @ file:///home/conda/feedstock_root/build_artifacts/confection_1740630664356/work
cryptography==46.0.5
cymem @ file:///Users/runner/miniforge3/conda-bld/cymem_1768533712950/work
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
ecdsa==0.19.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.135.1
filelock==3.25.0
flatbuffers==25.12.19
fsspec==2026.2.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hf-xet==1.3.2
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.5.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jinja2_1764517220/work
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kafka-python==2.3.0
kombu==5.6.2
kubernetes==35.0.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1754951200865/work
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1772445084013/work
mcp==1.26.0
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
mmh3==5.2.0
mpmath==1.3.0
murmurhash @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_murmurhash_1768531525/work
networkx==3.6.1
numpy @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_numpy_1770098395/work/dist/numpy-2.4.2-cp311-cp311-macosx_11_0_arm64.whl#sha256=4dd287d918a5abd30328a2f5bfa74bb734d8f9e7f12ce6a5675e4ae338b805c6
oauthlib==3.3.1
onnxruntime==1.24.2
openai==2.24.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
overrides==7.7.0
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079066/work
pluggy==1.6.0
posthog==5.4.0
preshed @ file:///Users/runner/miniforge3/conda-bld/preshed_1768546955091/work
prompt_toolkit==3.0.52
protobuf==6.33.5
psycopg2-binary==2.9.11
pyasn1==0.6.2
pybase64==1.4.3
pycparser==3.0
pydantic @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pydantic_1764434463/work
pydantic-settings==2.13.1
pydantic_core @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_pydantic-core_1762989051/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
PyJWT==2.11.0
PyMuPDF==1.27.1
PyPika==0.51.1
pyproject_hooks==1.2.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
redis==7.2.1
referencing==0.37.0
regex==2026.2.28
requests @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_requests_1766926974/work
requests-oauthlib==2.0.0
rich @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich_1771572730/work/dist
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
shellingham @ file:///home/conda/feedstock_root/build_artifacts/shellingham_1762858192781/work
six==1.17.0
smart_open @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_smart_open_1771856898/work/dist
sniffio==1.3.1
spacy @ file:///Users/runner/miniforge3/conda-bld/spacy_1763430381616/work
spacy-legacy @ file:///home/conda/feedstock_root/build_artifacts/spacy-legacy_1674550301837/work
spacy-loggers @ file:///home/conda/feedstock_root/build_artifacts/spacy-loggers_1694527114282/work
SQLAlchemy==2.0.48
srsly @ file:///Users/runner/miniforge3/conda-bld/srsly_1763419690452/work
sse-starlette==3.3.2
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
thinc @ file:///Users/runner/miniforge3/conda-bld/thinc_1763428532896/work
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_tqdm_1770153424/work
transformers==5.3.0
typer==0.24.0
typer-slim==0.24.0
typing-inspection @ file:///home/conda/feedstock_root/build_artifacts/typing-inspection_1764158251755/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.3
tzlocal==5.3.1
ujson @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_ujson_1756674083/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1767817748113/work
uvicorn==0.41.0
uvloop==0.22.1
vine==5.1.0
wasabi @ file:///home/conda/feedstock_root/build_artifacts/wasabi_1740630716360/work
watchfiles==1.1.1
wcwidth==0.6.0
weasel @ file:///home/conda/feedstock_root/build_artifacts/weasel_1763083786850/work
websocket-client==1.9.0
websockets==16.0
wrapt @ file:///Users/runner/miniforge3/conda-bld/wrapt_1770112014706/work
zipp==3.23.0
